build(core): Fix the build in the context of a git submodule
from TUG:
The gradle build file tries to infer the commit hash by reading the
files in the .git subdirectory. But in case the repo is checked out as
a submodule there is no such directory, only a file pointing to the
parent checkout.

Instead of depending on git internals just ask git directly.