chore(pdf-as-web): Handle error url not containing an explicit port
from TUG:
The servlet tries to sanitize the URL by parsing it and writing it back out.
In case the input URL doesn't contain an explicit port, URL.getPort() returns -1
which leads to "https://example.com:-1/mypath" in the template which isn't the
same origin as without a port and gets rejected by the browser.

Instead only add a port to the resulting URL if the input contains one as well.